+++ src/doom/p_switch.cpp
@@ -149,7 +149,7 @@ void P_InitSwitchList(void)
 	    alphSwitchList[i].episode;
 
 	// [crispy] remove MAXSWITCHES limit
-	if (slindex + 1 >= maxswitches)
+	if (slindex + 1 >= static_cast<int>(maxswitches))
 	{
 	    size_t newmax = maxswitches ? 2 * maxswitches : MAXSWITCHES;
 	    switchlist = static_cast<decltype(switchlist)>(I_Realloc(switchlist, newmax * sizeof(*switchlist)));
