+++ src/doom/r_things.cpp
@@ -443,7 +443,7 @@ void R_DrawMaskedColumn (column_t* column)
 	if (dc_yl <= mceilingclip[dc_x])
 	    dc_yl = mceilingclip[dc_x]+1;
 
-	if (static_cast<unsigned int>(dc_yl) <= dc_yh)
+	if (dc_yl <= dc_yh)
 	{
 	    dc_source = (byte *)column + 3;
 	    dc_texturemid = basetexturemid - (top<<FRACBITS);
