+++ src/doom/wi_stuff.cpp
@@ -422,7 +422,7 @@ void WI_drawLF(void)
     int y = WI_TITLEY;
 
     // [crispy] prevent crashes with maps without map title graphics lump
-    if (wbs->last >= num_lnames || lnames[wbs->last] == NULL)
+    if (wbs->last >= static_cast<int>(num_lnames) || lnames[wbs->last] == NULL)
     {
         V_DrawPatch((ORIGWIDTH - SHORT(finished->width)) / 2, y, finished);
         return;
@@ -466,7 +466,7 @@ void WI_drawEL(void)
     int y = WI_TITLEY;
 
     // [crispy] prevent crashes with maps without map title graphics lump
-    if (wbs->next >= num_lnames || lnames[wbs->next] == NULL)
+    if (wbs->next >= static_cast<int>(num_lnames) || lnames[wbs->next] == NULL)
     {
         return;
     }
